Ensure `tmp/` and `source/` directories exist
With the exception of `tmp/.gitkeep`, ensure that the `tmp/` directory
isn't included in source control.

Future commits will extend this project will output intermediate build
steps to the `tmp/`, so we don't want to persisted build artifacts to
our revision history.

As a complement to that, introduce and track `source/`, to ensure that
the project's source code will reside in that directory.