Elaborate on the project's goals in the README.md
Although the tool itself doesn't do anything yet, this commit introduces
the [`README.md`][README.md] file to serve as an informative welcome to
contributors and readers of the project's source code's history.

Once the project works as described, the [`README.md`][README.md] will
serve as the home page for the generated website.

[README.md]: #README.md