[GENERATED]: Scaffold the project with Middleman
The goals of the `git-read` project include transforming a project's
immutable (and static!) commit history into a website that tells the
story of the project under development.

The Middleman project is a tool to generate a website's static HTML and
asset files from a combination of source files, and outputs them to a
directory.

The following diff was generated by [Middleman][] as the output of
executing:

```bash
middleman init
```

[middleman]: https://middlemanapp.com/