Introduce Redcarpet Markdown Engine
Add a dependency on the [`redcarpet` gem][redcarpet] to parse the
project's imminent [Markdown][] files.

Additionally, configure Middleman to make use of the parser by:

* setting the `:markdown_engine` setting, along with additional
  `:markdown` configuration values from within [`config.rb`][config.rb]
* renaming the `source/index.html.erb` to `source/index.html.md`

The previous file extension sequence was significant, as it encoded a
[pre-processing sequence][templating] of [ERB] to HTML.

The _new_ sequence encodes that the source files will be written in
[Markdown][] (as denoted by the `.md`), then transformed into HTML (as
denoted by the `.html`).

To ensure that the templating engine change was successful, change the
test's [`assert_text`][assert_text] verification to a more strict
[`assert_selector`][assert_selector] that asserts that the `git-read`
text appears within an [`<h1>` element][h1], which is what Markdown
transforms text underlined with `===` into.

[redcarpet]: https://github.com/vmg/redcarpet/tree/v3.5.0#and-its-like-really-simple-to-use
[Markdown]: https://guides.github.com/features/mastering-markdown/
[config.rb]: #config.rb
[templating]: https://middlemanapp.com/basics/templating-language/#template-basics
[ERB]: https://ruby-doc.org/stdlib-2.7.0/libdoc/erb/rdoc/ERB.html
[assert_text]: https://rubydoc.info/github/teamcapybara/capybara/master/Capybara/Minitest/Assertions:assert_text
[assert_selector]: https://rubydoc.info/github/teamcapybara/capybara/master/Capybara/Minitest/Assertions:assert_selector
[h1]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements