Generate `/index.html` dynamically
Generate the contents of the site's `/index.html` page based on the
contents of the project's [`README.md`](#README.md).

To do so, this commit adds a dependency on the [`ruby-git`
gem][ruby-git], which introspects into the project to read the contents
of the _latest version_ of the file.

Since the contents of the [`README.md`](#README.md) are are dynamic, and
won't be declared statically in the `source/` directory, this commit
transitions the routing for the `/index.html` page into a [Middleman
`proxy` declaration][proxy], rendering the
[`source/README.html.md.erb`](#source/README.html.md.erb) template as
the `/index.html` page.

Once again, the sequencing of the template file's extensions are
crucial, as they hint to the pre-processor to first render ERB, then
yield those contents as static Markdown to be translated to HTML.

Set the page's [`<title>` element][mdn-title] to the first line of the
project's [`README.md`](#README.md).

Testing this behavior
---

The order that the directory, repository, and application are declared
and instantiated is also crucial.

Once the `Capybara.app = Middleman.server` line is invoked, a call to
`require` the application's [`config.rb`](#config.rb) file is made, and
`ruby-git`'s configuration is finalized.

To better control this, the test harness creates a temporary directory
with [`Dir.mktmpdir`][mktmpdir] from Ruby's Standard Library, which it
then passes to a `Git.init` call to prepare it as a Git directory.

Then, the new `with_git_repository` ensures that the `Middleman.server`
call is made after the environment has been configured to point the
application-side `Git` repository instance to the newly initialized
temporary `Git`-ready directory.

To coordinate this sequence, this commit introduces a dependency on
[`minitest-around` gem][minitest-around] for atomic test setup and
teardown, and the [`climate_control` gem][climate_control] for atomic
`ENV` modifications and roll-back.

[ruby-git]: https://github.com/ruby-git/ruby-git/tree/v1.6.0#examples
[proxy]: https://middlemanapp.com/advanced/dynamic-pages#defining-proxies
[mdn-title]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/title
[mktmpdir]: https://ruby-doc.org/stdlib-1.9.3/libdoc/tmpdir/rdoc/Dir.html#mktmpdir-method
[minitest-around]: https://github.com/splattael/minitest-around/tree/v0.5.0#unit-tests
[climate_control]: https://github.com/thoughtbot/climate_control/tree/v0.1.0