Add support for re-writing a project's history
If you're consuming this tool and using it to tell the story of your
project's origins and growth, you might be tempted to become a bit of a
revisionist historian.

This might come as a shock, but this project was not perfected
implemented from the start, and is in fact the result of countless
[`pick`, `squash`, `reword`, `fixup`, and `drop` rebase
operations][rebase].

While this can be harmless for pre-released projects, rewriting history
once the resulting website is deployed and publicized would result in
[`404`s and broken URLs][404].

To counteract this, `git-read` will include and compile pages for any
commits that are [`git tag`-ed][git-tag], along with the rest of their
history.

To denote that a commit is "out of date", and is no longer available on
the main branch, calls to Middleman's [`proxy`][proxy] have been updated
to pass in the `out_of_date` local Boolean to denote whether or not a
page is built from a tagged commit. When rendering from the main branch
or the project's `README.md`, `out_of_date` is declared as `false`.

In the case where the main branch and the tagged commit share ancestors,
the main branch's version will be rendered second, and won't display the
"stale" banner.

Test coverage
---

To ensure that the new feature is covered by tests, this commit extends
the new [`GitHelpers` module](#test/git_helpers.rb) to enable tests to
switch `git` branches (via `#switch_branch`), and to add tagged commits
(via `#add_tag`).

[rebase]: https://git-scm.com/book/en/v2/Git-Tools-Rewriting-History#_changing_multiple
[404]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/404
[git-tag]: https://git-scm.com/book/en/v2/Git-Basics-Tagging
[proxy]: https://middlemanapp.com/advanced/dynamic-pages#defining-proxies