Render how a commit changes the project's files
To provide the reader with ample context, render each commits' file
changes alongside their commit messages.

This commit introduces a wrapping
[`.container`][source/stylesheets/blocks/_container.scss]] CSS class, to
serve as a [flex box parent][display-flex].

Descendants will split the remaining page width amongst themselves.

In the case of the project root, the diff will be omitted, and the
commit message will span the width of the page (by virtue of it being
the [`:only-child` to the `.container`][only-child].
for project root

When a file's diff is longer than 100 lines, render the [`<details>`
element][details] as collapsed by default. To do so, omit the [`open`
attribute][open].

Likewise, when a commit's subject starts with `[GENERATED]`, don't
render the [`<details>` element][details] with an [`open`
attribute][open].

Determining the diff contents
---

When a commit has a parent, generating the diff with the [`ruby-git`
provided interface][ruby-git] is straightforward:

```ruby
diff = parent.diff(commit)
```

However, when the commit is the project's root, it has no parent, and
cannot be "diffed".

To counteract this restriction, this commit invokes a [`git
show`][git-show] command for each file included in the commit, passing
the [`--cc` combined diff flag][cc-flag]. The `--cc` flag will ensure
that the output contains the code diff. In order to separate out the
code from the prose, pass the command a [`--format=%n`
flag][format-flag] which includes the `%n` as a "newline" placeholder,
intended to serve as an omission of the prose.

The padding `\n` character is subsequently omitted by the trailing
Ruby-side call to [`String#strip`][strip].

The net result is that the contents of the diff include the changes to
the code, and _do not_ include the prose in the commit's message.

Test Coverage
---

To ensure that this new behavior is covered, this commit includes
corresponding System Tests.

To enable direct navigation to a commit, modify the existing `commit`
test helper to return the new commit object, so that the test can pass
it directly to the `commit_path` helper.

Desktop
---

![desktop view](https://user-images.githubusercontent.com/2575027/77105568-42179580-69f4-11ea-8bb0-17b51cf9be2a.png)

Mobile
---

![mobile view](https://user-images.githubusercontent.com/2575027/77105584-4479ef80-69f4-11ea-83a0-7d25d9daca07.png)

[display-flex]: https://developer.mozilla.org/en-US/docs/Web/CSS/display#display_flex
[overflow-x]: https://developer.mozilla.org/en-US/docs/Web/CSS/overflow-x
[only-child]: https://developer.mozilla.org/en-US/docs/Web/CSS/:only-child
[details]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details
[open]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details#attr-open
[ruby-git]: https://github.com/ruby-git/ruby-git/tree/v1.6.0#examples
[git-show]: https://git-scm.com/docs/git-show
[cc-flag]: https://git-scm.com/docs/git-show#_combined_diff_format
[format-flag]: https://git-scm.com/docs/pretty-formats
[strip]: https://ruby-doc.org/core-2.7.0/String.html#method-i-strip