Extract `Commit` model
Throughout the application, there are several occurrences of logic that
extract's a commit message's subject:

```ruby
commit.message.lines.first
```

In reaction to this repeated logic, we have several solutions at our
disposal.

Helper method
---

One potential solution is to extract `subject` method in the `helpers`
block in our [`config.rb`][config.rb]:

```ruby
helpers do
  def subject(commit)
    commit.message.lines.first
  end
end
```

Methods declared in the `helpers` block are made available within our
application templates (namely,
[`source/layouts/layout.erb`][source/layouts/layout.erb].

This would resolve the instances of duplication within that template,
but doesn't resolve instances within the [`config.rb`][config.rb]
itself.

Model Instance method
---

As an alternative to a helper method, this commit extracts the [`Commit`
class][lib/models/commit.rb], and declares it in the new `lib/models`
directory.

Similar to a helper method, an instance method centralizes the logic to
a single place. In the current implementation, the `Commit` class
inherits from [`SimpleDelegator`][SimpleDelegator], which makes the
methods declared in it an _expansion_ on the existing
[`Git::Object`][Git::Object] interface.

An additional benefit of an instance method on a model is that contexts
that have access to the `commit` template-local variable can invoke
`subject`, without requiring knowledge of the internals of the `commit`
reference. This style of message sending adheres to the ["Tell, Don't
Ask"][tell-dont-ask] principle of software design.

Centralizing the logic of transforming the [`Git::Object`
instances][Git::Object] into `git-read`-specific models will simplify
this project's future improvements.

[config.rb]: #config.rb
[source/layouts/layout.erb]: #source/layouts/layout.erb
[lib/models/commit.rb]: #lib/models/commit.rb
[SimpleDelegator]: https://ruby-doc.org/stdlib-2.5.1/libdoc/delegate/rdoc/SimpleDelegator.html
[Git::Object]: https://github.com/ruby-git/ruby-git/blob/v1.6.0/lib/git/object.rb
[tell-dont-ask]: https://pragprog.com/articles/tell-dont-ask