Add webpack for building JavaScript
This commit was influenced by the documentation in the [_Adding Webpack
to Middleman_][blog] article.

This commit introduces the [`webpack.config.js`][webpack.config.js]
file, declaring our application's entry point as the `site` key, along
with the `tmp/dist` directory as the destination to output assets to.

The `tmp/dist` directory declared in
[`webpack.config.js`][webpack.config.js] is echoed by the
[`config.rb`][config.rb] file's declaration of the `external_pipeline`
activation call's `source` property.

[blog]: https://dance.computer.dance/posts/2017/02/adding-webpack-to-middleman.html
[config.rb]: #config.rb
[webpack.config.js]: #webpack.config.js