Add `<details>` element polyfill
While support for the [`<details>` element][details] that powers this
application's commit history browser is widespread, [it is not supported
by all available browsers][compatibility].

Now that [NPM packages][npm] are available to the application's pages,
we can [polyfill][] missing support.

There is a [third-party, community sourced
polyfill][details-element-polyfill] for the [`<details>`
element][details] that this commit adds a dependency on.

[npm]: https://docs.npmjs.com/about-npm/
[details]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details
[compatibility]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details#Browser_compatibility
[polyfill]: https://developer.mozilla.org/en-US/docs/Glossary/Polyfill
[details-element-polyfill]: https://github.com/javan/details-element-polyfill/tree/2.4.0#html-details-element-polyfill