Navigate with Turbolinks
The HTML pages output by `git-read` commands are intended to be served
directly, through static site servers like [GitHub pages][gh-pages].

Despite the absence of a server-side component, the application's
navigation could still stand to benefit from the value that
[Turbolinks][] proposes:

> **Turbolinks® makes navigating your web application faster**. Get the
> performance benefits of a single-page application without the added
> complexity of a client-side JavaScript framework. Use HTML to render
> your views on the server side and link to pages as usual. When you
> follow a link, Turbolinks automatically fetches the page, swaps in its
> `<body>`, and merges its `<head>`, all without incurring the cost of a
> full page load.

To integrate with [Turbolinks][], modify existing event listeners that
are firing callbacks for the [`DOMContentLoaded`
event][DOMContentLoaded] declared on the [page's `document`][document],
replacing them with event listeners on the [`turbolinks:load`
event][turbolinks:load].

[gh-pages]: https://guides.github.com/features/pages/
[Turbolinks]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#observing-navigation-events
[DOMContentLoaded]: https://developer.mozilla.org/en-US/docs/Web/API/Document/DOMContentLoaded_event
[document]: https://developer.mozilla.org/en-US/docs/Web/API/Document
[turbolinks:load]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#observing-navigation-events

Scroll current Commit into view
---

When navigation to a commit completes, scroll that commit's node in the
history navigator into view.

When the browser finished loading and parsing the page's necessary
assets and elements, it fires a [`DOMContentLoaded`
event][DOMContentLoaded] on the [`document` object][document].

This commit adds an [event listener][] that fires a callback whenever a
`DOMContentLoaded` event is triggered.

When the browser finalizes loading and parsing its contents, scroll any
descendant elements annotated with the [`[aria-current="page"]`
attribute][aria-current] into view my invoking the [`.scrollIntoView()`
method][scrollIntoView] directly on the elements.

Unfortunately, support for the [`Element.scrollIntoView()`
method][scrollIntoView] is not as widespread as support for the
`<details>` element. Luckily, there is a [community-sourced
polyfill][polyfill] for invoking [`.scrollIntoView()` with
options][options].

Invoking `.scrollIntoView()` upon `DOMContentLoaded` will improve the
visibility of any `[aria-current]` elements present on the page.
However, pages include a current element in the project's history that
is hidden within a `<details>` element that does not specify an [`[open]`
attribute][details-open].

To account for that, extend the `DOMContentLoaded` callback so that it
queries all on [`<details> elements`][details] that are present on the
page and attaches callbacks that will be triggered when those [elements’
`[open]` attribute][details-open] is [`toggle`-d][details-toggle].

Whenever a [`<details>` element][details] is expanded, scroll any
descendant elements annotated with the [`[aria-current="page"]`
attribute][aria-current] into view my invoking the [`.scrollIntoView()`
method][scrollIntoView] directly on the element.

[DOMContentLoaded]: https://developer.mozilla.org/en-US/docs/Web/API/Document/DOMContentLoaded_event
[document]: https://developer.mozilla.org/en-US/docs/Web/API/Document
[event listener]: https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener
[queries]: https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelectorAll
[details]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details
[details-open]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details#attr-open
[details-toggle]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details#Events
[aria-current]: https://tink.uk/using-the-aria-current-attribute/
[scrollIntoView]: https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollIntoView
[polyfill]: https://github.com/wessberg/scroll-behavior-polyfill/tree/v2.0.13#applying-the-polyfill
[options]: https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollIntoView#Parameters