Extract SVG images
Several affordances throughout the application are visualized through
[`<svg>` elements][mdn-svg]. Prior to this commit, they were declared
directly in the template HTML.

Some of the application's iconography was shared across several parts of
the interface. For instance, the commit history's `<summary>` element in
mobile dimensions shared the same icon `<svg>` elements as the file
diff's `<summary>` element.

To re-use the shared elements, this commit adds a dependency on the
[`middleman-inline_svg` gem][middleman-inline_svg], which reads files
from the `source/images` directory and renders the contents of files
inline as HTML, similar to what was done before.

The main value that `middleman-inline_svg` provides is that it affords
an opportunity for the rendering context to insert HTML attributes
(like [`class`][class] and [`role`][role]) or an SVG [`<title>`
element][svg-title].

To that end, this commit also [configures the extension][configuration]
to specify [`role="img"`][role-img] for each SVG, since each collection
of elements represents a single entity.

[mdn-svg]: https://developer.mozilla.org/en-US/docs/Web/SVG/Element/svg
[middleman-inline_svg]: https://github.com/thoughtbot/middleman-inline_svg#usage
[class]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/class
[role]: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles
[svg-title]: https://developer.mozilla.org/en-US/docs/Web/SVG/Element/title
[configuration]: https://github.com/thoughtbot/middleman-inline_svg#configuration
[role-img]: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/Role_Img e54f7c1d6a7877353a5d88777b914bf47600f4ec