Install Tailwind
Add an npm dependency on Tailwind and install it according to
[TailwindCSS Installation: Add Tailwind as a PostCSS
plugin][tailwind-postcss].

Tailwind is designed to participate with the JavaScript build ecosystem,
and integrates directly into a Webpack-enabled pipeline like the one
we've set up previously. To do so, we'll also need to [pull in several
loaders](#package.json) and their corresponding [webpack
plugins](#webpack.config.js).

Replace the Middelman-provided `autoprefixer` integration with the
[npm-provided version][autoprefixer].

[autoprefixer]: https://github.com/postcss/autoprefixer
[tailwind-postcss]: https://tailwindcss.com/docs/installation#add-tailwind-as-a-post-css-plugin