Syntax highlight Markdown fenced code blocks
The [syntax-highlighting styles][syntax-css] are wholly generated from
the [`rouge` gem's Github theme][rouge].

To enable it, [configure Middleman's Markdown rendering
engine][syntax-markdown] by adding a call to `activate :syntax` to the
project's [`config.rb`](#config.rb).

To ensure that `middleman-syntax` and the `@tailwindcss/typography`
styles are not at odds with one another, customize
`@tailwindcss/typography` to omit some of its predefined `<pre>` element
styles.

Append the [syntax highlighting][syntax] CSS to the bottom of
[`source/stylesheets/syntax.css.erb`](#source/stylesheets/syntax.css.erb),
and extend the Tailwind `.prose` classes to account for the new
`.highlight` class.

[syntax-css]: https://github.com/middleman/middleman-syntax/tree/v3.0.0#css
[syntax-markdown]: https://github.com/middleman/middleman-syntax/tree/v3.0.0#markdown
[syntax]: https://github.com/middleman/middleman-syntax/tree/v3.0.0#configuration
[rouge]: https://github.com/rouge-ruby/rouge/blob/v3.17.0/lib/rouge/themes/github.rb